requirements.txt
Django==4.2.2
djangorestframework==3.14.0
joblib==1.3.1
librosa==0.10.0.post2
numpy==1.24.4
pandas==2.0.3
requests==2.31.0
scikit-learn==1.2.2
soundfile==0.12.1
gunicorn==20.1.0
django-environ==0.10.0
dj-database-url==2.0.0
psycopg2-binary==2.9.6
whitenoise[brotli]==6.5.0
